Keep older player stats.
Make the `leaderboard_display` view keep track of all players ever,
and add a `leaderboard_display_recent` view to show only the players who
played in the last two weeks.

This have the effect to make the `playerStats` function return stats
even for players that didn't play in the last two weeks, so their score
is kept in the leaderboard when they start playing again.

Closes #50.